requirements.txt
djangorestframework
django
Pygments
Markdown
coreapi
psycopg2-binary
dj-database-url
gunicorn
whitenoise
PyYAML

inflection
